requirements.txt
passlib
flask
mysql-connector-python
flask-cors
flask-socketio
dotenv